Move some more functions to detect-utils
Keep moving generic functions out from detect to detect-utils.
This is also in preparation of refactoring detect-system.
