+++ hardware/tests/test_detect_utils.py
@@ -101,3 +101,127 @@ class TestDetectUtils(unittest.TestCase):
 
     def test_get_cidr(self):
         self.assertEqual(detect_utils.get_cidr('255.255.0.0'), '16')
+
+    def test_fix_bad_serial_zero(self):
+        hwl = [('system', 'product', 'serial', '0000000000')]
+        detect_utils.fix_bad_serial(hwl, 'uuid', '', '')
+        self.assertEqual(detect_utils.get_value(
+            hwl, 'system', 'product', 'serial'), 'uuid')
+
+    def test_fix_bad_serial_mobo(self):
+        hwl = [('system', 'product', 'serial', '0123456789')]
+        detect_utils.fix_bad_serial(hwl, '', 'mobo', '')
+        self.assertEqual(detect_utils.get_value(
+            hwl, 'system', 'product', 'serial'), 'mobo')
+
+    @mock.patch('hardware.detect_utils.from_file', side_effect=IOError())
+    @mock.patch('hardware.detect_utils.output_lines',
+                side_effect=[
+                    sample('lscpu').split('\n'),
+                    sample('lscpux').split('\n')])
+    def test_get_cpus(self, mock_output_lines, mock_throws_ioerror):
+        hw = []
+        detect_utils.get_cpus(hw)
+        self.assertEqual(hw, detect_utils_results.GET_CPUS_RESULT)
+        calls = []
+        calls.append(mock.call('/sys/devices/system/cpu/smt/control'))
+        # Once per socket
+        for i in range(2):
+            calls.append(mock.call('/sys/devices/system/cpu/cpufreq/boost'))
+        # Once per processor
+        for i in range(1):
+            calls.append(mock.call(('/sys/devices/system/cpu/cpufreq/'
+                                    'policy{}/scaling_governor'.format(i))))
+            calls.append(mock.call(('/sys/devices/system/cpu/cpu{}/cpufreq/'
+                                    'scaling_governor'.format(i))))
+        # NOTE(tonyb): We can't use assert_has_calls() because it's too
+        # permissive.  We want an exact match
+        self.assertEqual(calls, mock_throws_ioerror.mock_calls)
+
+    @mock.patch('hardware.detect_utils.from_file', side_effect=IOError())
+    @mock.patch('hardware.detect_utils.output_lines',
+                side_effect=[
+                    sample('lscpu-7302').split('\n'),
+                    sample('lscpu-7302x').split('\n')])
+    def test_get_cpus_7302(self, mock_output_lines, mock_throws_ioerror):
+        self.maxDiff = None
+        hw = []
+        detect_utils.get_cpus(hw)
+        self.assertEqual(hw, detect_utils_results.GET_CPUS_7302_RESULT)
+
+    @mock.patch('hardware.detect_utils.from_file', side_effect=IOError())
+    @mock.patch('hardware.detect_utils.output_lines',
+                side_effect=[
+                    sample('lscpu-vm').split('\n'),
+                    sample('lscpu-vmx').split('\n'),
+                    ('powersave',),
+                    ('powersave',)])
+    def test_get_cpus_vm(self, mock_output_lines, mock_throws_ioerror):
+        hw = []
+        detect_utils.get_cpus(hw)
+        self.assertEqual(hw, detect_utils_results.GET_CPUS_VM_RESULT)
+        calls = []
+        calls.append(mock.call('/sys/devices/system/cpu/smt/control'))
+        # Once per socket
+        for i in range(1):
+            calls.append(mock.call('/sys/devices/system/cpu/cpufreq/boost'))
+        # Once per processor
+        for i in range(2):
+            calls.append(mock.call(('/sys/devices/system/cpu/cpufreq/'
+                                    'policy{}/scaling_governor'.format(i))))
+            calls.append(mock.call(('/sys/devices/system/cpu/cpu{}/cpufreq/'
+                                    'scaling_governor'.format(i))))
+        # NOTE(tonyb): We can't use assert_has_calls() because it's too
+        # permissive.  We want an exact match
+        self.assertEqual(calls, mock_throws_ioerror.mock_calls)
+
+    @mock.patch('hardware.detect_utils.from_file', side_effect=IOError())
+    @mock.patch('hardware.detect_utils.output_lines',
+                side_effect=[
+                    sample('lscpu_aarch64').split('\n'),
+                    sample('lscpux_aarch64').split('\n'),
+                    ('powersave',),
+                    ('powersave',)])
+    def test_get_cpus_aarch64(self, mock_output_lines, mock_throws_ioerror):
+        self.maxDiff = None
+        hw = []
+        detect_utils.get_cpus(hw)
+        self.assertEqual(hw, detect_utils_results.GET_CPUS_AARCH64_RESULT)
+        calls = []
+        calls.append(mock.call('/sys/devices/system/cpu/smt/control'))
+        # Once per socket
+        for i in range(4):
+            calls.append(mock.call('/sys/devices/system/cpu/cpufreq/boost'))
+        # Once per processor
+        for i in range(8):
+            calls.append(mock.call(('/sys/devices/system/cpu/cpufreq/'
+                                    'policy{}/scaling_governor'.format(i))))
+            calls.append(mock.call(('/sys/devices/system/cpu/cpu{}/cpufreq/'
+                                    'scaling_governor'.format(i))))
+        # NOTE(tonyb): We can't use assert_has_calls() because it's too
+        # permissive.  We want an exact match
+        self.assertEqual(calls, mock_throws_ioerror.mock_calls)
+
+    @mock.patch('hardware.detect_utils.from_file', side_effect=IOError())
+    @mock.patch('hardware.detect_utils.output_lines',
+                side_effect=[
+                    sample('lscpu_ppc64le').split('\n'),
+                    sample('lscpux_ppc64le').split('\n')])
+    def test_get_cpus_ppc64le(self, mock_output_lines, mock_throws_ioerror):
+        hw = []
+        detect_utils.get_cpus(hw)
+        self.assertEqual(hw, detect_utils_results.GET_CPUS_PPC64LE)
+        calls = []
+        calls.append(mock.call('/sys/devices/system/cpu/smt/control'))
+        # Once per socket
+        for i in range(2):
+            calls.append(mock.call('/sys/devices/system/cpu/cpufreq/boost'))
+        # Once per processor
+        for i in range(144):
+            calls.append(mock.call(('/sys/devices/system/cpu/cpufreq/'
+                                    'policy{}/scaling_governor'.format(i))))
+            calls.append(mock.call(('/sys/devices/system/cpu/cpu{}/cpufreq/'
+                                    'scaling_governor'.format(i))))
+        # NOTE(tonyb): We can't use assert_has_calls() because it's too
+        # permissive.  We want an exact match
+        self.assertEqual(calls, mock_throws_ioerror.mock_calls)
